+++ syndicate/core/resources/lambda_resource.py
@@ -317,6 +317,14 @@ class LambdaResource(BaseResource):
         waiter = self.lambda_conn.get_waiter('function_exists')
         waiter.wait(FunctionName=name)
 
+        if meta.get('max_retries') is not None:
+            _LOG.debug('Setting lambda event invoke config')
+            invoke_config = self.lambda_conn.put_function_event_invoke_config(
+                function_name=name,
+                max_retries=meta.get('max_retries')
+            )
+            _LOG.debug(f'Created lambda invoke config: {invoke_config}')
+
         log_group_name = name
         retention = meta.get('logs_expiration')
         if retention:
@@ -474,6 +482,27 @@ class LambdaResource(BaseResource):
         waiter = self.lambda_conn.get_waiter('function_updated_v2')
         waiter.wait(FunctionName=name)
         _LOG.info(f'Waiting has finished')
+
+        if meta.get('max_retries') is not None:
+            try:
+                _LOG.debug('Updating lambda event invoke config')
+                invoke_config = \
+                    self.lambda_conn.update_function_event_invoke_config(
+                        function_name=name,
+                        max_retries=meta.get('max_retries')
+                    )
+            except ClientError as e:
+                if e.response['Error']['Code'] == 'ResourceNotFoundException':
+                    _LOG.debug('Lambda event invoke config is absent. '
+                               'Creating a new one')
+                    invoke_config = \
+                        self.lambda_conn.put_function_event_invoke_config(
+                            function_name=name,
+                            max_retries=meta.get('max_retries')
+                        )
+                else:
+                    raise e
+            _LOG.debug(invoke_config)
 
         response = self.lambda_conn.get_function(name)
         _LOG.info(f'Lambda describe result: {response}')