+++ CHANGELOG.md
@@ -4,15 +4,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.6.2] - 2023-11-14
+
+# [1.8.0] - 2023-12-12
+### Added
+- A new Maven plugin goal 'assemble-lambda-layer-files' has been added to the 'deployment-configuration-processor' plugin
+- A new Java lambda example with assembling layer files and url config has been added
+- An example of adding custom SDK to layer and url config for lambda has been added to Java examples folder
+### Changed
+- The deployment-configuration-processor version bumped to 1.10.0
+
+# [1.7.5] - 2023-12-11
+- Fixed API Gateway service integration credentials building
+
+# [1.7.4] - 2023-12-08
+- Fixed lambda event invoke config creating and updating
+
+# [1.7.3] - 2023-12-07
+- Fixed lambda deployment in extended prefix mode
+
+# [1.7.2] - 2023-11-14
 - Fixed SNS subscription deletion with SNS topic deletion
 - Fixed processing of retries for `InvalidParameterValueException`
 
-# [1.6.1] - 2023-11-14
+# [1.7.1] - 2023-11-14
 - Added waiting for IAM role creation
 
-# [1.6.0] - 2023-11-10
+# [1.7.0] - 2023-11-10
 - Implemented extended prefix mode
+
+# [1.6.0] - 2023-12-07
+- Added `max_retries` attribute setting for creating and updating 
+lambda configuration
 
 # [1.5.0] - 2023-11-08
 ### Added